Repair Tests

+++ src/test/coins_tests.cpp
@@ -378,7 +378,7 @@ BOOST_AUTO_TEST_CASE(updatecoins_simulation_test)
 
             // Call UpdateCoins on the top cache
             CTxUndo undo;
-            UpdateCoins(tx, *(stack.back()), undo, height);
+            UpdateCoins(tx, *(stack.back()), undo, height, uint256());
 
             // Update the utxo set for future spends
             utxoset.insert(outpoint);
